Move code back specific to cobra
