+++ cmd/origin.go
@@ -24,7 +24,6 @@ import (
 
 	"github.com/pelicanplatform/pelican/config"
 	"github.com/pelicanplatform/pelican/metrics"
-	"github.com/pelicanplatform/pelican/utils"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -84,13 +83,13 @@ https://jwt.io/introduction.
 Additional profiles that expand on JWT are supported. They include scitokens2 and
 wlcg. For more information about these profiles, see https://scitokens.org/technical_docs/Claims
 and https://github.com/WLCG-AuthZ-WG/common-jwt-profile/blob/master/profile.md, respectively`,
-		RunE: utils.CliTokenCreate,
+		RunE: cliTokenCreate,
 	}
 
 	originTokenVerifyCmd = &cobra.Command{
 		Use:   "verify",
 		Short: "Verify a Pelican origin token",
-		RunE:  utils.VerifyToken,
+		RunE:  verifyToken,
 	}
 )
